Fix csma timeout

+++ sys/include/net/ieee802154/submac.h
@@ -196,7 +196,7 @@ struct ieee802154_submac {
     ieee802154_csma_be_t be;            /**< CSMA-CA backoff exponent params */
     bool wait_for_ack;                  /**< SubMAC is waiting for an ACK frame */
     uint32_t ack_timeout_us;            /**< ACK timeout in µs */
-    uint16_t csma_backoff_us;           /**< CSMA sender backoff period in µs */
+    uint32_t csma_backoff_us;           /**< CSMA sender backoff period in µs */
     uint16_t panid;                     /**< IEEE 802.15.4 PAN ID */
     uint16_t channel_num;               /**< IEEE 802.15.4 channel number */
     uint8_t channel_page;               /**< IEEE 802.15.4 channel page */
